feat: align RM integration conventions with mediamtx
