+++ Dockerfile
@@ -6,6 +6,7 @@ RUN npm run build
 
 # Dockerfile for Node.js server
 FROM node:22-slim AS production
+COPY --from=pvarki/kw_product_init:latest /kw_product_init /kw_product_init
 RUN apt-get update && apt-get install -y \
         curl \
     && rm -rf /var/lib/apt/lists/* \
@@ -27,8 +28,9 @@ COPY --from=build /usr/src/app/dist ./public
 # Expose the port the app runs on
 EXPOSE 3000
 
-COPY ./entrypoint.sh /
-RUN chmod +x /entrypoint.sh
+COPY ./entrypoint.sh /entrypoint.sh
+COPY ./container-init.sh /container-init.sh
+RUN chmod +x /entrypoint.sh /container-init.sh
 
 ENTRYPOINT ["/entrypoint.sh"]
