+++ backend/config/index.js
@@ -1,4 +1,5 @@
 import dotenv from 'dotenv';
+import { getManifestRmCertCn } from '../utils/kraftwerkManifest.js';
 
 dotenv.config();
 
@@ -10,13 +11,15 @@ const parseBoolean = (value, defaultValue = false) => {
     return normalized === '1' || normalized === 'true' || normalized === 'yes' || normalized === 'on';
 };
 
+const manifestRmCertCn = getManifestRmCertCn();
+
 const config = {
     baseUrl: process.env.BASE_URL || '/',
     port: process.env.PORT || 3000,
     databaseUrl: process.env.DATABASE_URL,
     rmMtlsEnforce: parseBoolean(process.env.RM_MTLS_ENFORCE, false),
     rmMtlsHeader: process.env.RM_MTLS_HEADER || 'x-clientcert-dn',
-    rmExpectedCertCn: process.env.RM_EXPECTED_CERT_CN || 'rasenmaeher',
+    rmExpectedCertCn: process.env.RM_EXPECTED_CERT_CN || manifestRmCertCn || 'rasenmaeher',
 };
 
 export default config;
