+++ entrypoint.sh
@@ -1,8 +1,14 @@
-#!/bin/bash
+#!/bin/bash -l
 
 set -euo pipefail
 
-echo "Running migrations"
-npm run migrate:up
-echo "Starting server"
-npm run start
+. /container-init.sh
+
+if [ "$#" -eq 0 ]; then
+    echo "Running migrations"
+    npm run migrate:up
+    echo "Starting server"
+    exec npm run start
+fi
+
+exec "$@"
